Add a option for null values to be accepted in validations
This change allows for a single field to be null (in addition to
its other types). This is different than required in that you
can actively change a value to None instead of omitting or
ignoring it. It is essentially the ignore_none_values, allowing
for more fine grained control down to the field level.