Make Validator less picky about exact data types...
Python is quite duck-typey, so it seems bad to say that certain elements of
the document being checked should be exact instances of things like dict
and list. Rather, they should conform to the more general types
MutableMapping and MutableSquence, which then allows for validation of
more complex entities as long as they adhere to the interfaces of the
built-in dict and list types.