Validate schemas before they are used to validate documents...
This throws out the same errors as we currently do at document-parsing time,
but allows us to fail in cases where the document doesn't reference pieces of
the schema which are illegal.

This is important should there be a rarely-used part of the schema which is
incorret, as we will fail earlier, hopefully reducing debug times!

See issue #41 for full details of development discussion.