Add an simpler 'list of types' example.
The original example might suggest that you need to define a 'schema'
rule when using a list of types, which is only true if you have a type
which actually requires a schema.

Hopefully, adding an even simpler example will clarify that a 'schema'
is not mandatory in order to use the 'list of types' feature.