Refactoring concerning DefinitionSchema
the purpose of these changes was to let DefinitionSchema behave like a proper dictionary
to allow safe runtime-manipulation of a schema.
if such manipulation raises an exception that is handled, the previous state is retained.

while working on it, i stumbled over some issues that are addressed.

to have more descriptive and unambigious names:

- renames ValidationError to DocumentError
- renames ERROR_SCHEMA_TYPE to ERROR_SCHEMA_TYPE_TYPE
  - ERROR_SCHEMA_CONSTRAINT_TYPE_TYPE would be more correct
- renames ERROR_SCHEMA_FORMAT to ERROR_SCHEMA_TYPE
- adds a ERROR_SCHEMA_CONSTRAINT_TYPE for constraints that are not defined as mapping

other:

- exceptions are inherited from Exception, so they are not caught by `except ValueError`