Overhauls error system
This is supposed to be the first step towards a more versatile error-reporting.

While validating errors are just collected in a list, not inserted into a tree.
The errors are now represented by ValitationError-instances that keep various
aspects of an error.

The former error representation is now achieved by a BasicErrorHandler and still
the default that returns from `Validator.errors`.

All tests have been adapted and succeed. Custom validators are still functional.