Catch exception if the distribution is not found
When the distribution of Cerberus cannot be found, catch the exception
instead of raising an error and exiting the program. This allows to
consume the package without installing it, e.g. as a git submodule. In
the exception case the __version__ string is set to `unknown`.

Fixes #471

(cherry picked from commit 7896257ab2cb2c6b1e3e59f120254ec181cabfeb)