Create FUNDING.yml
This will add a Sponsor button on the repo page. In due time (a couple weeks if it gets approved) we will be able to also GH Sponsors, which supports multiple GH users (hey funkyfuture! -if interested). Take a look at eve repo for a gist of how it works.